# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AfroHit Studio API
  description: |
    Public REST contract. All endpoints under `/api/v1` require a Clerk bearer JWT.
    The Studio Chat endpoint accepts free-form prompts and orchestrates the labs by
    calling internal tools (hooks, lyrics, beats, vocals, etc.).
  version: 0.1.0
servers:
  - url: http://localhost:4000
    description: local
  - url: https://api.afrohit.studio
    description: production
security:
  - bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/projects/{projectId}/lyrics/generate:
    post:
      summary: Generate full lyrics around an approved hook
      parameters: [{ in: path, name: projectId, required: true, schema: { type: string } }]
      responses: { '201': { description: created } }
  /api/v1/projects/{projectId}/lyrics/{lyricId}/approve:
    post:
      summary: Approve a lyric draft
      parameters:
        - { in: path, name: projectId, required: true, schema: { type: string } }
        - { in: path, name: lyricId, required: true, schema: { type: string } }
      responses: { '200': { description: ok } }
...
  /api/v1/projects/{projectId}/mixes:
    post:
      summary: Create a mix with a preset
      parameters: [{ in: path, name: projectId, required: true, schema: { type: string } }]
      responses: { '202': { description: queued } }
  /api/v1/projects/{projectId}/mixes/master:
    post:
      summary: Master a mix
      parameters: [{ in: path, name: projectId, required: true, schema: { type: string } }]
      responses: { '202': { description: queued } }
...